update the Music component
